Fix github-linguist statistics (#547)
* Added .gitattributes

Tell git that the test/testdata/ files and Jupyter Notebooks are binary
files, ie. do not display diffs and do not touch line endings in those
files.

* Moved 'CQ examples.ipynb' to examples directory

Now github-linguist will classify it as documentation, and the massive
line count in that file will stop skewing the language statistics for
CadQuery on Github.