untilNextFace and untilLastFace for extrude and cutblind
Added new options for extrude and cutBlind
- untilNextFace : extrude/cut until the next face found along the wire normal dir
- untilLastFace : same operation but until the last face found

I quickly tested and both options for both methods seems to work.
I need to clean up the code a bit, I had to add quite a lot of if/else statement that may be better written.

It has been a been tricky to decide what lines should go to what level so feel free to tell me if you woud have done it in a different manner