Adds ability to define polygons with a circumscribed circle
This adds the option to define polygons by a circumscribed circle
in addition to the previous of inscribed circles. Inscribing remains
the default. Circumscribing is useful in certain circumstances
such as when the distance from the center of the polygon to the
midpoint of one of the segments needs to be defined, e.g. when
tiling the polygons along the x-axis.