added docs on the CQ different APIs
As invoked with https://github.com/CadQuery/cadquery/issues/909

This PR aims to bring more documentation about the different CadQuery APIs and how one can take advantage of them.

This has been written with my own understanding prism so it is rather subjective.  I'm not sure these concepts have been properly named before so this is an attempt to clarify things for eveybody