Update svg.py
The exportSVG function produced incorectly scaled svgs. Also the UnitOfMeasure is not something that should be guessed based on a heuristic but something that should be passed along by the user with the default being mm. The UnitScale has been set to 3.7795 as per the svg spec and creates svgs with the correct size.