feat: Add viewup parameter to show in vis.py
This allows the user to setup the ViewUp position of the camera.
The default is with Y up (0, 1, 0) but for projects like creating
shapes for 3D printing Z up is more conventional. With this change
it is difficult to manage.

The default is Y up but that default can now be overridden by passing
`viewup=(0, 0, 1)`