Fix additional spaces in Python code blocks in tests (#1861)
The correct alignment is the 3 column, aligned to the `code-block` directive,
not the 4 column.

The extra space is not removed by the docutils parser, so the docs
all had an extra space in them, which is very annoying for copy-pasting,
especially into limited editors like the CQ-editor where you have to clean up
every code sample before it works (otherwise you get IndentationError).