Update quickstart documentation inconsistencies (#1871)
The referenced line numbers following the code-block in the More Holes section weren't pointing at the highlighted lines which I believe was not the intent, I updated it to what seemed correct.

I also re-arranged some sections to follow the top-down flow the rest of the document seems to be following.

I'm new this library and the CAD domain entirely so these stood out to my beginner's eyes. Really liking this project so far!